make reflector optional again
That the reflector is mandatory argument is a breaking change and is not necessary. The problem with this change is that you can not just add a `new JitReflector()` in a library where you don't know if the final version is working with angular 4.2 or angular 4.1. You just can't import the JitReflector as is did not exist in 4.1.